+++ tests/test_loggers.py
@@ -6,6 +6,8 @@ from __future__ import absolute_import, division, print_function
 
 import sys
 
+import pytest
+
 from structlog._compat import StringIO
 from structlog._loggers import (
     PrintLogger,
@@ -14,11 +16,15 @@ from structlog._loggers import (
     ReturnLoggerFactory,
     WRITE_LOCKS,
 )
+from structlog.stdlib import _NAME_TO_LEVEL
 
 
 def test_return_logger():
     obj = ['hello']
     assert obj is ReturnLogger().msg(obj)
+
+
+STDLIB_MSG_METHODS = [m for m in _NAME_TO_LEVEL if m != 'notset']
 
 
 class TestPrintLogger(object):
@@ -61,6 +67,15 @@ class TestPrintLogger(object):
         PrintLogger(sio)
         assert sio in WRITE_LOCKS
 
+    @pytest.mark.parametrize("method", STDLIB_MSG_METHODS)
+    def test_stdlib_methods_support(self, method):
+        """
+        PrintLogger implements methods of stdlib loggers.
+        """
+        sio = StringIO()
+        getattr(PrintLogger(sio), method)('hello')
+        assert 'hello' in sio.getvalue()
+
 
 class TestPrintLoggerFactory(object):
     def test_does_not_cache(self):
@@ -83,6 +98,16 @@ class TestPrintLoggerFactory(object):
         the factory, they are not passed to the logger.
         """
         PrintLoggerFactory()(1, 2, 3)
+
+
+class ReturnLoggerTest(object):
+    @pytest.mark.parametrize("method", STDLIB_MSG_METHODS)
+    def test_stdlib_methods_support(self, method):
+        """
+        ReturnLogger implements methods of stdlib loggers.
+        """
+        v = getattr(ReturnLogger(), method)('hello')
+        assert 'hello' == v
 
 
 class TestReturnLoggerFactory(object):